- minor mods to support/build.sh so that command generator will be run ONLY if the gradle process exits normally - Excel: handles OLE2NotOfficeXmlFileException when attempting to open a XLS file as a Excel-2007 format - Nexial Filter: reordered NexialFilterComparator to improve REGEX search & group - adding org.junit.runners.ParentRunner as another class to signify current environment as "unit testing" (and thus disable console interaction) - fixed failed tests

+++ support/build.sh
@@ -8,13 +8,19 @@ else
 	gradle clean testJar installDist
 fi
 
-mkdir -p build/install/nexial-core/support
-cp -Rvf support/nexial*.* build/install/nexial-core/support
+build_ret=$?
 
-# generate the latest command listing
-cd build/install/nexial-core/support
-chmod -fR 755 *.sh
-./nexial-command-generator.sh
+if [[ ${build_ret} = 0 ]] ; then
+	mkdir -p build/install/nexial-core/support
+	cp -Rvf support/nexial*.* build/install/nexial-core/support
 
-cd ..
-rm -frv support
+	# generate the latest command listing
+	cd build/install/nexial-core/support
+	chmod -fR 755 *.sh
+	./nexial-command-generator.sh
+
+	cd ..
+	rm -frv support
+else
+	echo gradle build failed!
+fi
